backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.0
celery==5.3.4
redis==5.0.1
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1

# Linting & Formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
